+++ packages/dialect-react-ui/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [UNRELEASED]
 
+## [1.0.0-beta.22] - 2022-07-26
 - feature: configurable notifications in `<NotificationsButton/>` component settings
 - feature: configurable notifications in `<BroadcastForm/>` component
 - fix: types for react 18